src/main/java/streams/SumBigDecimals.java
package streams;

import java.math.BigDecimal;
import java.util.stream.Stream;

public class SumBigDecimals {

    public BigDecimal sumFirstN_asDoubles(int n) {
        double total = Stream.iterate(BigDecimal.ONE, bd -> bd.add(BigDecimal.ONE))
                .limit(n)
                .mapToDouble(BigDecimal::doubleValue)
                .sum();
        return new BigDecimal(total + "");
    }

    public BigDecimal sumFirstN_usingReduce(int n) {
        return Stream.iterate(BigDecimal.ONE, bd -> bd.add(BigDecimal.ONE))
                .limit(n)
                .reduce(BigDecimal::add).orElse(BigDecimal.ZERO);
    }

    public BigDecimal sumFirstN_usingReduce_verbose(int n) {
        return Stream.iterate(BigDecimal.ONE, bd -> bd.add(BigDecimal.ONE))
                .limit(n)
                .reduce((bigDecimal, augend) -> {
                    System.out.println("accumulator = " + bigDecimal + ", arg = " + augend);
                    return bigDecimal.add(augend);
                }).orElse(BigDecimal.ZERO);
    }

    // Off by one error, because 1 is never doubled
    public BigDecimal sumDoubles(int n) {
        BigDecimal two = new BigDecimal("2");
        return Stream.iterate(BigDecimal.ONE, bd -> bd.add(BigDecimal.ONE))
                .limit(n)
                .reduce((total, e) -> {
                    System.out.println("total = " + total + ", e = " + e);
                    return total.add(e.multiply(two));
                }).orElse(BigDecimal.ZERO);
    }

    public BigDecimal sumDoublesInitialized(int n) {
        BigDecimal two = new BigDecimal("2");
        return Stream.iterate(BigDecimal.ONE, bd -> bd.add(BigDecimal.ONE))
                .limit(n)
                .reduce(BigDecimal.ZERO, (total, e) -> {
                    System.out.println("total = " + total + ", e = " + e);
                    return total.add(e.multiply(two));
                });
    }

    public double productOfDoubles(int n) {
        return Stream.iterate(1.0, k -> k + 1)
                .limit(n)
                .reduce((acc, val) -> {
                    System.out.println("acc = " + acc + ", val = " + val);
                    return acc * val;
                }).orElse(0.0);
    }

    public double productOfDoublesWithIdentity(int n) {
        return Stream.iterate(1.0, k -> k + 1)
                .limit(n)
                .reduce(1.0, (acc, val) -> acc * val);
    }
}
